requirement.txt
django>=3.1
django-import-export>=2.7.1
